refactor(rust cycle 19): admin_ops/hermes_handlers 372→dir{chat 102, runs 140, run 118, types 18}
